// src/app/shared/interfaces/article.interface.ts
import { ProfileInterface } from './profile.interface'

export interface ArticleInterface {
  author: ProfileInterface
  body: string
  comments: string[]
  createdAt: string
  description: string
  favorited: boolean
  favoritesCount: number
  slug: string
  tagList: string[]
  title: string
  updatedAt: string
}